pyproject.toml
[project]
name = "shop-drf"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-autoslug>=1.9.9",
    "django-filter>=25.1",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "load-dotenv>=0.1.0",
    "pillow>=11.2.1",
    "psycopg>=3.2.6",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.6",
]
docs = [
    "drf-spectacular>=0.28.0",
]
